Add end switch and calibration support
Add support for two end switches, the header pin numbers of which get passed
into the PyStepper constructor. The first switch (pin) in the array is used for
calibration and must be placed such that it detects the limit in the negatie
direction. Functionality of the upper limit switch has not been tested yet.