Never return a JSON array as an AJAX response
For security reasons, the response to any AJAX request must never be a
plain JSON array. Although security isn't an issue here, apply this
principle in preparation for using the Tornado server, which explicitly
forbids returning plain arrays.

This also fixes the ugliness of hard-coding the value labels (keys) for
the status values in the server. This was initially done in an attempt
to keep the status object flexible, so one could add additional
variables without having to modify the html. It's not clear that this is
needed. Returning a plain object seems like the better approach.