TestServer: Post move data as form, rather than JSON
The previous implementation was clearly wrong, as it was posting with a
content type of 'application/x-www-form-urlencoded', but sending plain
JSON in the body. At least the type and actual content should match.

This change switches to posting urlencoded forms, rather than JSON.
Although the latter is a valid option, posting as plain forms has the
advantage of allowing posts from an actual HTML form, without any
JavaScript code on the browser side.