Add TornadoServer as alternative to TestServer
In preparation for delivering status (position) updates through a
Websocket, rather than REST polling, re-implement the TestServer using
the Tornado library. This takes care of much of the crud and handles
errors much more gracefully than the hand-knit TestServer.

In order to use this, the tornado package needs to be installed:
        $ pip install tornado

Due to the single-threaded, asynchronous nature of Tornado, this is a
lot snappier than Python's standard BaseHTTPServer and hopefully results
in fewer hiccups during stepper movements.