TornadoServer: Add SockJS/Websocket support
Rather than making an AJAX query every second, create a SockJS
connection when first loading the page, then stream status messages over
that connection. This significantly reduces communication overhead on
both sides.

While most current browsers support Websockets out of the box, use the
SockJS library instead. Not only does this minimize the chances of
browser compatibility issues, but it's also a preparation for
integrating the PyStepper with Octoprint, which uses Tornado/SockJS as
well.

This change adds a dependency on the sockjs-tornado Python package.
Moreover, because the latest sockjs-tornado is incompatible with the
latest tornado package, we need to downgrade the latter:

        $ pip install tornado==4.0.2
        $ pip install sockjs-tornado

In order to keep TestServer functional, the JavaScript code now checks
the response headers and only attempts a SockJS connection if the
'Server' header starts with 'Tornado'. Otherwise, it falls back to
polling the status.