Add 'calibrated' state to stepper daemon
Make the stepper daemon remember whether it has been calibrated and
prohibit absolute positioning as long as calibration was not performed.

Also reflect this in the web page and disable the 'move' input and
button as long as calibration has not yet been performed. Relative
movements are allowed at any time, as is calibration, which is just a
special form of a relative movement.